Option for PowerDnsProvider to request notification after changes
This patch introduces `notify` option to PowerDnsProvider, which
instructs the provider to ask the pdns server for sending a
notification (DNS NOTIFY) to its secondary servers after changing zone
content.

The option is off by default since the notification request will fail
unless the pdns server operates as a primary for the zone.